requirements.txt
Flask~=0.12
flask-cors
jsonschema
pyyaml
py4j>=0.10.6
spylon>=0.3.0
unicodecsv>=0.14.1
